implementação do kafka e correção em algumas validações

+++ desafio-hinova/src/main/java/com/demo/desafio_hinova/Repository/UsuarioRepository.java
@@ -3,5 +3,6 @@ package com.demo.desafio_hinova.Repository;
 import com.demo.desafio_hinova.Model.Usuarios;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuarioRepository extends JpaRepository<Usuarios, Integer> {
+
+public interface UsuarioRepository extends JpaRepository<Usuarios, Long> {
 }
